api/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
_tflow_select=2.3.0=mkl
abseil-cpp=20211102.0=hd4dd3e8_0
absl-py=1.4.0=py39h06a4308_0
aiohttp=3.9.3=py39h5eee18b_0
aiosignal=1.2.0=pyhd3eb1b0_0
anyio=4.2.0=py39h06a4308_0
astunparse=1.6.3=py_0
async-timeout=4.0.3=py39h06a4308_0
attrs=23.1.0=py39h06a4308_0
blas=1.0=openblas
blinker=1.6.2=py39h06a4308_0
brotli-python=1.0.9=py39h6a678d5_7
bzip2=1.0.8=h5eee18b_5
c-ares=1.19.1=h5eee18b_0
ca-certificates=2024.3.11=h06a4308_0
cachetools=4.2.2=pyhd3eb1b0_0
certifi=2024.2.2=py39h06a4308_0
cffi=1.15.1=py39h74dc2b5_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.7=py39h06a4308_0
cryptography=41.0.3=py39h130f0dd_0
exceptiongroup=1.2.0=py39h06a4308_0
expat=2.5.0=h6a678d5_0
fastapi=0.103.0=py39h06a4308_0
flatbuffers=24.3.25=pypi_0
freetype=2.12.1=h4a9f257_0
frozenlist=1.4.0=py39h5eee18b_0
gast=0.4.0=pyhd3eb1b0_0
giflib=5.2.1=h5eee18b_3
google-auth=2.29.0=pypi_0
google-auth-oauthlib=1.2.0=pypi_0
google-pasta=0.2.0=pyhd3eb1b0_0
grpc-cpp=1.46.1=h33aed49_1
grpcio=1.62.1=pypi_0
h11=0.14.0=py39h06a4308_0
h5py=3.9.0=py39he06866b_0
hdf5=1.12.1=h70be1eb_2
icu=58.2=he6710b0_3
idna=3.4=py39h06a4308_0
importlib-metadata=7.0.1=py39h06a4308_0
intel-openmp=2022.1.0=h9e868ea_3769
jpeg=9e=h5eee18b_1
keras=2.15.0=pypi_0
keras-preprocessing=1.1.2=pyhd3eb1b0_0
krb5=1.20.1=h568e23c_1
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libclang=18.1.1=pypi_0
libcurl=8.2.1=h91b91d3_0
libdeflate=1.17=h5eee18b_1
libedit=3.1.20230828=h5eee18b_0
libev=4.33=h7f8727e_1
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libnghttp2=1.52.0=ha637b67_1
libopenblas=0.3.21=h043d6bf_0
libpng=1.6.39=h5eee18b_0
libprotobuf=3.20.3=he621ea3_0
libssh2=1.10.0=h37d81fd_2
libstdcxx-ng=11.2.0=h1234567_1
libtiff=4.5.1=h6a678d5_0
libuuid=1.41.5=h5eee18b_0
libwebp=1.3.2=h11a3e52_0
libwebp-base=1.3.2=h5eee18b_0
lz4-c=1.9.4=h6a678d5_0
markdown=3.4.1=py39h06a4308_0
markupsafe=2.1.3=py39h5eee18b_0
mkl=2020.2=256
mkl-service=2.3.0=py39he8ac12f_0
ml-dtypes=0.2.0=pypi_0
multidict=6.0.4=py39h5eee18b_0
ncurses=6.4=h6a678d5_0
numpy=1.26.3=py39heeff2f4_0
numpy-base=1.26.3=py39h8a23956_0
oauthlib=3.2.2=py39h06a4308_0
openjpeg=2.4.0=h3ad879b_0
openssl=1.1.1w=h7f8727e_0
opt_einsum=3.3.0=pyhd3eb1b0_1
packaging=23.1=py39h06a4308_0
pillow=10.0.1=py39ha6cbd5a_0
pip=23.3.1=py39h06a4308_0
protobuf=3.20.3=py39h6a678d5_0
pyasn1=0.4.8=pyhd3eb1b0_0
pyasn1-modules=0.2.8=py_0
pycparser=2.21=pyhd3eb1b0_0
pydantic=1.10.12=py39h5eee18b_1
pyjwt=2.4.0=py39h06a4308_0
pyopenssl=23.2.0=py39h06a4308_0
pysocks=1.7.1=py39h06a4308_0
python=3.9.0=hdb3f193_2
python-multipart=0.0.6=py39h06a4308_0
re2=2022.04.01=h295c915_0
readline=8.2=h5eee18b_0
requests=2.31.0=py39h06a4308_1
requests-oauthlib=1.3.0=py_0
rsa=4.7.2=pyhd3eb1b0_1
scipy=1.11.4=py39heeff2f4_0
setuptools=68.2.2=py39h06a4308_0
six=1.16.0=pyhd3eb1b0_1
snappy=1.1.10=h6a678d5_1
sniffio=1.3.0=py39h06a4308_0
sqlite=3.41.2=h5eee18b_0
starlette=0.27.0=py39h06a4308_0
tbb=2021.8.0=hdb19cb5_0
tensorboard=2.15.2=pypi_0
tensorboard-data-server=0.7.2=pypi_0
tensorboard-plugin-wit=1.8.1=py39h06a4308_0
tensorflow=2.15.0=pypi_0
tensorflow-estimator=2.15.0=pypi_0
tensorflow-io-gcs-filesystem=0.36.0=pypi_0
termcolor=2.1.0=py39h06a4308_0
tk=8.6.12=h1ccaba5_0
typing-extensions=4.9.0=py39h06a4308_1
typing_extensions=4.9.0=py39h06a4308_1
tzdata=2024a=h04d1e81_0
urllib3=2.1.0=py39h06a4308_1
uvicorn=0.20.0=py39h06a4308_0
werkzeug=2.3.8=py39h06a4308_0
wheel=0.41.2=py39h06a4308_0
wrapt=1.14.1=py39h5eee18b_0
xz=5.4.6=h5eee18b_0
yarl=1.9.3=py39h5eee18b_0
zipp=3.17.0=py39h06a4308_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.5=hc292b87_0
